+++ api/src/web/mod.rs
@@ -10,23 +10,36 @@
  * its affiliates is strictly prohibited.
  */
 
+use std::env;
 use std::sync::Arc;
+use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
 use askama::Template;
-use axum::extract::{Path as AxumPath, State as AxumState};
-use axum::middleware::Next;
-use axum::response::{Html, IntoResponse, Response};
-use axum::routing::{get, post, Router};
+use axum::{
+    extract::{Host, Path as AxumPath, State as AxumState},
+    middleware::Next,
+    response::{Html, IntoResponse, Redirect, Response},
+    routing::{get, post, Router},
+    Extension,
+};
+use axum_extra::extract::cookie::{Cookie, Key, PrivateCookieJar};
 use base64::prelude::*;
 use hyper::http::header::{CONTENT_TYPE, WWW_AUTHENTICATE};
-use hyper::http::{Request, StatusCode};
+use hyper::http::{HeaderMap, Request, StatusCode};
+use itertools::Itertools;
+use oauth2::basic::BasicClient;
+use oauth2::{
+    AuthUrl, ClientId, ClientSecret, CsrfToken, PkceCodeChallenge, RedirectUrl, Scope, TokenUrl,
+};
 use rpc::forge::forge_server::Forge;
 use rpc::forge::{self as forgerpc};
 use tower_http::normalize_path::NormalizePath;
 
 use crate::api::Api;
 use crate::cfg::CarbideConfig;
+use crate::CarbideError;
 
+mod auth;
 mod domain;
 mod dpu_versions;
 mod expected_machine;
@@ -47,12 +60,130 @@ mod vpc;
 
 const WEB_AUTH: &str = "admin:Welcome123";
 
+const AUTH_TYPE_ENV: &str = "CARBIDE_WEB_AUTH_TYPE";
+const AUTH_CALLBACK_ROOT: &str = "auth-callback";
+
+// Details https://entra.microsoft.com/#view/Microsoft_AAD_RegisteredApps/ApplicationMenuBlade/~/Overview/appId/5ae5fa35-be8e-44cc-be7b-01ff76af5315/isMSAApp~/false
+const OAUTH2_AUTH_ENDPOINT_ENV: &str = "CARBIDE_WEB_OAUTH2_AUTH_ENDPOINT";
+const DEFAULT_OAUTH2_AUTH_ENDPOINT: &str =
+    "https://login.microsoftonline.com/43083d15-7273-40c1-b7db-39efd9ccc17a/oauth2/v2.0/authorize";
+
+const OAUTH2_TOKEN_ENDPOINT_ENV: &str = "CARBIDE_WEB_OAUTH2_TOKEN_ENDPOINT";
+const DEFAULT_OAUTH2_TOKEN_ENDPOINT: &str =
+    "https://login.microsoftonline.com/43083d15-7273-40c1-b7db-39efd9ccc17a/oauth2/v2.0/token";
+
+const CARBIDE_WEB_PRIVATE_COOKIEJAR_KEY_ENV: &str = "CARBIDE_WEB_PRIVATE_COOKIEJAR_KEY";
+const CARBIDE_WEB_HOSTNAME_ENV: &str = "CARBIDE_WEB_HOSTNAME";
+
+const OAUTH2_CLIENT_SECRET_ENV: &str = "CARBIDE_WEB_OAUTH2_CLIENT_SECRET";
+const OAUTH2_CLIENT_ID_ENV: &str = "CARBIDE_WEB_OAUTH2_CLIENT_ID";
+const DEFAULT_OAUTH2_CLIENT_ID: &str = "5ae5fa35-be8e-44cc-be7b-01ff76af5315";
+
+const ALLOWED_ACCESS_GROUPS_LIST_ENV: &str = "CARBIDE_WEB_ALLOWED_ACCESS_GROUPS";
+const DEFAULT_ALLOWED_ACCESS_GROUPS_LIST: &str = "swngc-forge-admins,ngc-forge-sre,swngc-forge-dev";
+
+const ALLOWED_ACCESS_GROUPS_ID_LIST_ENV: &str = "CARBIDE_WEB_ALLOWED_ACCESS_GROUPS_ID_LIST";
+const DEFAULT_ALLOWED_ACCESS_GROUPS_ID_LIST: &str =
+    "1f13d1bb-6d7e-4fa5-9abf-93e24e7b5a4e,80f709a0-77a7-4a15-899d-7abba0ffdc1f,d03b7e2a-673b-4088-9af0-545a2d2f4c5d";
+
 const SORTABLE_JS: &str = include_str!("../../templates/static/sortable.min.js");
 const SORTABLE_CSS: &str = include_str!("../../templates/static/sortable.min.css");
 
+#[derive(Clone)]
+pub(crate) struct Oauth2Layer {
+    client: BasicClient,
+    http_client: reqwest::Client,
+    private_cookiejar_key: Key,
+    allowed_access_groups_filter: String,
+    allowed_access_groups_ids: Vec<String>,
+}
+
 /// All the URLs in the admin interface. Nested under /admin in api.rs.
-pub fn routes(api: Arc<Api>) -> NormalizePath<Router> {
-    NormalizePath::trim_trailing_slash(
+pub fn routes(api: Arc<Api>) -> eyre::Result<NormalizePath<Router>> {
+    // Just something to let us transition more easily.
+    // By default, everything will be the original basic-auth,
+    // so we can deploy this all over and flip on azure auth with
+    // some env-vars.  When everything is switched over, we can
+    // clean this up this up, maybe directly send the struct without the option wrapper,
+    // and only ever use oauth2 if we want.
+    let oauth_extension_layer = match env::var(AUTH_TYPE_ENV)
+        .unwrap_or("basic".to_string())
+        .to_lowercase()
+        .as_str()
+    {
+        "oauth2" => {
+            // Get our cookiejar key so we can add it as an extension.
+            let private_cookiejar_key = Key::try_from(
+                env::var(CARBIDE_WEB_PRIVATE_COOKIEJAR_KEY_ENV)
+                    .map_err(|e| {
+                        CarbideError::GenericError(format!(
+                            "{}: {}",
+                            CARBIDE_WEB_PRIVATE_COOKIEJAR_KEY_ENV, e
+                        ))
+                    })?
+                    .as_bytes(),
+            )?;
+
+            // Grab the details for which groups are allowed to access the web UI.
+            let allowed_access_groups_ids = env::var(ALLOWED_ACCESS_GROUPS_ID_LIST_ENV)
+                .unwrap_or(DEFAULT_ALLOWED_ACCESS_GROUPS_ID_LIST.to_string())
+                .split(",")
+                .map(|s| s.to_lowercase())
+                .collect();
+
+            let allowed_access_groups_filter = env::var(ALLOWED_ACCESS_GROUPS_LIST_ENV)
+                .unwrap_or(DEFAULT_ALLOWED_ACCESS_GROUPS_LIST.to_string())
+                .split(",")
+                .map(|s| format!("\"displayName:{}\"", s.to_lowercase()))
+                .join(" OR ");
+
+            // Build the  OAuth2 client.
+            let client = BasicClient::new(
+                ClientId::new(
+                    env::var(OAUTH2_CLIENT_ID_ENV).unwrap_or(DEFAULT_OAUTH2_CLIENT_ID.to_string()),
+                ),
+                Some(ClientSecret::new(
+                    env::var(OAUTH2_CLIENT_SECRET_ENV).map_err(|e| {
+                        CarbideError::GenericError(format!("{}: {}", OAUTH2_CLIENT_SECRET_ENV, e))
+                    })?,
+                )),
+                AuthUrl::new(
+                    env::var(OAUTH2_AUTH_ENDPOINT_ENV)
+                        .unwrap_or(DEFAULT_OAUTH2_AUTH_ENDPOINT.to_string()),
+                )?,
+                Some(TokenUrl::new(
+                    env::var(OAUTH2_TOKEN_ENDPOINT_ENV)
+                        .unwrap_or(DEFAULT_OAUTH2_TOKEN_ENDPOINT.to_string()),
+                )?),
+            )
+            .set_redirect_uri(RedirectUrl::new(format!(
+                "https://{}/admin/{}",
+                env::var(CARBIDE_WEB_HOSTNAME_ENV).unwrap_or("localhost:1079".to_string()),
+                AUTH_CALLBACK_ROOT,
+            ))?);
+
+            let http_client = {
+                let builder = reqwest::Client::builder();
+                let builder = builder
+                    .redirect(reqwest::redirect::Policy::none())
+                    .connect_timeout(Duration::new(5, 0)) // Limit connections to 5 seconds
+                    .timeout(Duration::new(15, 0)); // Limit the overall request to 15 seconds
+
+                builder.build()?
+            };
+
+            Some(Oauth2Layer {
+                client,
+                private_cookiejar_key,
+                allowed_access_groups_filter,
+                allowed_access_groups_ids,
+                http_client,
+            })
+        }
+        _ => None,
+    };
+
+    Ok(NormalizePath::trim_trailing_slash(
         Router::new()
             .route("/", get(root))
             .route("/static/:filename", get(static_data))
@@ -147,9 +278,103 @@ pub fn routes(api: Arc<Api>) -> NormalizePath<Router> {
             .route("/vpc.json", get(vpc::show_json))
             .route("/vpc/:vpc_id", get(vpc::detail))
             .route("/search", get(search::find))
-            .layer(axum::middleware::from_fn(auth_basic))
+            .route(&format!("/{AUTH_CALLBACK_ROOT}"), get(auth::callback))
+            .layer(axum::middleware::from_fn(auth_oauth2))
+            .layer(Extension(oauth_extension_layer))
             .with_state(api),
+    ))
+}
+
+pub async fn auth_oauth2<T>(
+    Host(hostname): Host,
+    headers: HeaderMap,
+    req: Request<T>,
+    next: Next<T>,
+) -> Result<Response, StatusCode> {
+    let oauth_extension_layer = match req.extensions().get::<Option<Oauth2Layer>>() {
+        None => {
+            tracing::error!("failed to find oauth2 extension layer");
+            return Err(StatusCode::INTERNAL_SERVER_ERROR);
+        }
+        Some(o) => match o {
+            None => {
+                return auth_basic(req, next).await;
+            }
+            Some(oa) => oa.to_owned(),
+        },
+    };
+
+    // /auth-callback should pass through because that's
+    // where microsoft will call back after the auth attempt.
+    if req.uri().path().starts_with("/auth-callback") {
+        return Ok(next.run(req).await);
+    }
+
+    let cookiejar: PrivateCookieJar = PrivateCookieJar::from_headers(
+        &headers,
+        oauth_extension_layer.private_cookiejar_key.clone(),
+    );
+
+    // If it exists, do we still want to accept it?
+    if let Some(c) = cookiejar.get("sid").map(|cookie| cookie.value().to_owned()) {
+        if let Ok(expiraton_timestamp) = c.parse::<u64>() {
+            let now_seconds = SystemTime::now()
+                .duration_since(UNIX_EPOCH)
+                .map_err(|e| {
+                    tracing::error!(%e, "failed to get system time for oauth2 expiration check");
+                    StatusCode::INTERNAL_SERVER_ERROR
+                })?
+                .as_secs();
+
+            // Still valid?  Let'em pass through to where they wanted
+            // to go.
+            if now_seconds < expiraton_timestamp {
+                return Ok(next.run(req).await);
+            }
+        }
+    }
+
+    // If not found or expired, we'll grab the oauth client and redirect to Azure for auth.
+
+    // Generate a PKCE challenge.
+    let (pkce_challenge, pkce_verifier) = PkceCodeChallenge::new_random_sha256();
+
+    // Generate the full authorization URL.
+    let (auth_url, csrf_state) = oauth_extension_layer
+        .client
+        .authorize_url(CsrfToken::new_random)
+        .add_scope(Scope::new("User.Read".to_string()))
+        .set_pkce_challenge(pkce_challenge)
+        .url();
+
+    // Store the pkce verifier so we can use it later
+    // during code exchange when they hit our callback URL.
+    // Using this with a cookie is a little weird, but it'll be encrypted.
+    let pkce_cookie = Cookie::build("pkce_verifier", pkce_verifier.secret().to_owned())
+        .domain(hostname.clone())
+        .path("/")
+        .secure(true)
+        .http_only(true)
+        .finish();
+
+    // Store the csrf state so we can compare the state we get back from Azure
+    // when they hit our callback URL.
+    let csrf_cookie = Cookie::build("csrf_state", csrf_state.secret().to_owned())
+        .domain(hostname)
+        .path("/")
+        .secure(true)
+        .http_only(true)
+        .finish();
+
+    Ok((
+        cookiejar
+            .remove(csrf_cookie.clone())
+            .remove(pkce_cookie.clone())
+            .add(csrf_cookie)
+            .add(pkce_cookie),
+        Redirect::to(auth_url.as_ref()),
     )
+        .into_response())
 }
 
 pub async fn auth_basic<T>(req: Request<T>, next: Next<T>) -> Result<Response, StatusCode> {
@@ -163,7 +388,7 @@ pub async fn auth_basic<T>(req: Request<T>, next: Next<T>) -> Result<Response, S
         }
         Some(auth_val) => {
             let Ok(auth_val) = auth_val.to_str() else {
-                tracing::debug!("Invalid auth header");
+                tracing::error!("Invalid auth header");
                 return Err(StatusCode::BAD_REQUEST);
             };
             if !is_valid_auth(auth_val) {